:recycle: Create Feature Module for Single-file Downloads
